+++ internal/server/application.go
@@ -1,10 +1,13 @@
 package server
 
 import (
+	"context"
+
 	"github.com/rs/zerolog"
 
 	"github.com/agentstation/starmap"
 	"github.com/agentstation/starmap/pkg/catalogs"
+	pkgsync "github.com/agentstation/starmap/pkg/sync"
 )
 
 // Application is the catalog and operational role consumed by the HTTP server.
@@ -13,5 +16,7 @@ type Application interface {
 	CatalogState() (starmap.CatalogState, error)
 	Readiness() (starmap.CatalogReadiness, error)
 	Starmap(...starmap.Option) (*starmap.Client, error)
+	Sync(context.Context, ...pkgsync.Option) (*pkgsync.Result, error)
+	UpdatesEnabled() bool
 	Logger() *zerolog.Logger
 }