+++ internal/server/handlers/application.go
@@ -1,8 +1,11 @@
 package handlers
 
 import (
+	"context"
+
 	"github.com/agentstation/starmap"
 	"github.com/agentstation/starmap/pkg/catalogs"
+	pkgsync "github.com/agentstation/starmap/pkg/sync"
 )
 
 type application interface {
@@ -10,4 +13,5 @@ type application interface {
 	CatalogState() (starmap.CatalogState, error)
 	Readiness() (starmap.CatalogReadiness, error)
 	Starmap(...starmap.Option) (*starmap.Client, error)
+	Sync(context.Context, ...pkgsync.Option) (*pkgsync.Result, error)
 }