+++ pkg/catalogs/provider.go
@@ -213,9 +213,9 @@ type ProviderPrivacyPolicy struct {
 
 // ProviderRetentionPolicy represents how long data is kept and deletion practices.
 type ProviderRetentionPolicy struct {
-	Type     ProviderRetentionType `json:"type" yaml:"type"`                             // Type of retention policy
-	Duration *time.Duration        `json:"duration,omitempty" yaml:"duration,omitempty"` // nil = forever, 0 = immediate deletion
-	Details  *string               `json:"details,omitempty" yaml:"details,omitempty"`   // Human-readable description
+	Type     ProviderRetentionType `json:"type" yaml:"type"`                                                   // Type of retention policy
+	Duration *time.Duration        `json:"duration,omitempty" yaml:"duration,omitempty" swaggertype:"integer"` // nil = forever, 0 = immediate deletion
+	Details  *string               `json:"details,omitempty" yaml:"details,omitempty"`                         // Human-readable description
 }
 
 // ProviderGovernancePolicy represents oversight and moderation practices.