+++ server/generate.go
@@ -1,0 +1,3 @@
+package server
+
+//go:generate gomarkdoc -e -o README.md . --repository.url https://github.com/agentstation/starmap --repository.default-branch main --repository.path /server